changed instances of uint to uint256

+++ contracts/token/PausableToken.sol
@@ -11,11 +11,11 @@ import '../lifecycle/Pausable.sol';
 
 contract PausableToken is StandardToken, Pausable {
 
-  function transfer(address _to, uint _value) whenNotPaused returns (bool) {
+  function transfer(address _to, uint256 _value) whenNotPaused returns (bool) {
     return super.transfer(_to, _value);
   }
 
-  function transferFrom(address _from, address _to, uint _value) whenNotPaused returns (bool) {
+  function transferFrom(address _from, address _to, uint256 _value) whenNotPaused returns (bool) {
     return super.transferFrom(_from, _to, _value);
   }
 }
