+++ test/helpers/DayLimitMock.sol
@@ -2,17 +2,17 @@ pragma solidity ^0.4.11;
 import "../../contracts/DayLimit.sol";
 
 contract DayLimitMock is DayLimit {
-  uint public totalSpending;
+  uint256 public totalSpending;
 
-  function DayLimitMock(uint _value) DayLimit(_value) {
+  function DayLimitMock(uint256 _value) DayLimit(_value) {
     totalSpending = 0;
   }
 
-  function attemptSpend(uint _value) external limitedDaily(_value) {
+  function attemptSpend(uint256 _value) external limitedDaily(_value) {
     totalSpending += _value;
   }
 
-  function setDailyLimit(uint _newLimit) external {
+  function setDailyLimit(uint256 _newLimit) external {
     _setDailyLimit(_newLimit);
   }
